Permission-testing works correctly when permissions lack display names.
The code in the UI that checks permissions starts out by fetching the
perms from the server, and making a map that records which permission
names are provided. Previously, it mapped permission names to the
corresponding human-readable display names. This worked fine for
testing whether a user has a permission by seeing whether the map has
the relevant entry. But when the displayName is undefined, as it is in
the mod-users test data (as opposed to the auth-blackbox VM), this
fails -- permissions are present, but undefined.

But since the permission-checking never actually _uses_ the
displayName, we now fix this by simply mapping from permission name to
the constant `true`.

Fixes STRIPES-296.