Inconsequential change in hasPerm API
Previously, stripes.hasPerm would return false if the user's
permissions have not yet been loaded, true if they have an the
specified permission is present, and undefined if the permissions have
been loaded by their specified one is not among them.

Now the sense of undefined and false are reversed: undefined means
"we can't tell whether or not the user will have the permission,
because the permissions have not yet been loaded"; and false means "we
have the permissions, and this isn't one of them".

This is clearer and more honest; but it make no difference at all to
client code that merely checks "if stripes.hasPerm(perm)". The purpose
of the change is only to make logging more informative.