Move list of permissions below "Log out" in user menu
It still only appears when the showAuth configuration item is true, of
course.

The reason for this change is that now the list of permissions can be
so long that the "Log out" option was getting pushed off the bottom of
the page, so it was impossible to log out!

Also: the introductory text "Perms:" is now bolded.