Pull ../../stripes-connect/doc/thinking-in-stripes.md into dev-guide.md
At present there is rather more duplication between these two
documents than we want. And since "Thinking in Stripes" is something
that developers need to do, it's very natural for that document to
become a section of the broader Guide.

I have not yet deduplicated material, or moved it around. Some of what
is presently near the start of the thinking-in-stripes section needs
to be near to top of the whole document.

Part of STRIPES-103 and STRIPES-254.