Add "Note on sharing resources between components"
This document is now "complete" in the sense that every section I
thought of and sketched out is now written. No doubt there are many
more sections that _could_ be included -- I'll look forward to
feedback on that.

This basically completes STRIPES-254.

(But its various subissues live on, specifically STRIPES-103 and
STRIPES-281; and indeed STRIPES-256.)