New utility component <AddContext> puts stripes object on context
I wanted this to be completely generic, so that you can use

        <AddContext context={{ foo: 1, bar: 'baz' }}>
          <SomeComponentThatUsesFooAndBar />
        </AddContext>

but that doesn't _quite_ work because you need to specify in the
static childContextTypes object what properties you're going to supply
on the context. I don't know if there's a way around that, but it's
late and that is a problem for another day.

Provides the mechanism for STRIPES-388.