Navigation to any app continues from last visited page
We do this just by noticing every time we navigate to a page within an
app, and making note that that app's most recently visited page is as
seen; then the link that we follow to get to the app next time is
taken from the note.

Fixes the URL part of STRIPES-362.