Key bindings are now loaded rom mod-configuration
Since this can only be done after login (we need the authentication
token to access mod-configuration), the structure of the code has
changed, and discoverKeyBindings.js has gone completely.

This seems to be working code, but actual bindings are not yet being
activated consistently in ui-* modules that try to use them, so there
may be bugfixes yet to come.

Towards STRPCORE-2.