Generate IDs that should be palatable to Okapi
Instead of "@folio/users-2.7.1-SNAPSHOT.324", "folio_users-2.7.1":
* Leading "@" is discarded
* "/" in module-name is replaced with "_"
* Everything after "-" in version-number is discarded.

I think this a bad idea: we are throwing information away here. We
actively want to know if Okapi is being given Snapshot 324. But for
now, at least, we'll take the path of least resistance.

Fixes STRPCORE-12.

(Unless Okapi also objects to underscores in IDs, in which case we'll
need a further tweaks.)