New script, util/link-stripes
After cloning the modules of a Stripes development environment --
consisting of one or more of the stripes-* and ui-* git modules from
https://github.com/folio-org/ -- run this script in the directory that
contains them to set up yarn linking. The first time you run it, use
the "-i" option to initialise the low-level repositories as yarn-link
destinations. Thereafter it can be omitted.

Fixes STRPCORE-14.