Remove unused nav buttons
The top navigation bar is populated via the automatically discovered
list of applications. However, in addittion to those, there is a list
of six navigation buttons that don't appear to serve any purpose other
than to demonstrate the visual style. For someone using the system
this can be confusing since it immediately raises the question "should
this do something?", or "should it go somewhere?"

This removes those questions by removing those buttons.