Allow pluggable history interface in stripes router.
Popular test runners like [Karam][1] and [Testem][2] allow you to run
your test suite against multiple browsers. In order to do this, they
invert control so that a browser is driving your test suite, and not
that the test suite is driving a browser.

However, this creates a problem in certain scenarios, because in order
for the browser to "capture" an application, it can't have the URL
_actually_ changing as the test navigates around the application. so in
order to use one of these runners with react router, it needs to be
using an instance of 'MemoryRouter' and not 'BrowserRouter'. That way,
the URL is changing from the perspective of the application, but not
from the perspective of the actual browser.

This allows the `Root` component to accept a `history` property, which
should be an instance of [History][3], so that inside a browser test
harness, you can drive your stripes application with a MemoryHistory.

To accomplish this, we use the prop types mechanism, supplying the
current history (browser history) as the default.

> Note: in order to add some structurual sanity checking for the
> history object, we check for the `length`, `action`, `push`, and
> `replace` properties. This isn't a complete list, but should prevent
> something that is clearly bogus from getting passed in.

[1]: https://karma-runner.github.io/1.0/index.html
[2]: https://github.com/testem/testem
[3]: https://github.com/reacttraining/history