[STCOR-46] Make Front page component pluggable.
The landing page of a stripes application is the first thing that
a user sees when they navigate to it. At the moment, there is a
placeholder text there, but you can't customize that text.

This change adds a new type of plugin "frontpage" which, if present
will be presented to the user.

Implementation Notes
====================

- if a plugged-in component is rendered in a stripes connected
  environment, the `Pluggable` component will attempt to connect the
  component that it wraps. However, if we're rendering outside the
  context of an app (as in the case of the frontpage), then the
  `connect()` method is not present on the stripes context. This adds
  a default `connect()` which passes through the original component
  without any decoration

- The `Front` component needed to be extended because it now needs to
  setup a context in order to support pluggability. We should consider
  having a default context that is set up at the very top of the
  render tree.