Move root component reducers reference to be instance specific
If the Root component were to be mounted and unmounted any reducers
previously added would remain in the reducers object and so new
reducers with the same key would never be overridden.

This can be problematic during testing when you want a new app (and store)
instance for each test case. In those instances the new store reducers
are never added to the new store because their keys would be present
in the already existing reducers hash.

This change simply moves the reducers reference to be instance
specific. So when new instances of the Root component are created the
reducers hash is set to it's initial value.