New RootWithIntl component provides intl to Stripes object
The main rendering that was in `<Root>` is now pushed down into
`<RootWithIntl>`, since the latter's `render` method, having been
invoked inside '<IntlProvider>`, is able to provide an enhanced clone
of the Stripes object that has in the `intl` object added.

This fixes STCOR-52, though I will shuffle the code around a bit
before I finally close that one.