Allow for contextual epics
As an application container, Stripes cannot know ahead of time which
epics will be installed into its Redux middleware. Because of that, if
apps or extensions are going to use epics, then they will need to have
a way to add them from within their own lifecycle. E.g. when the
application component mounts, it can communicate which epics it
contains.

Furthermore, it is a possibility that a stripes application contained
by the platform will be mounted and unmounted several times, and so
the extension mechanism must not "overinstall" epics as the user
navigates to and from an application so that they end up being run in
duplicate, triplicate, etc...

This change contains two enhancements. First, the epics Redux
middleware was removed from being held as a static constant in the
`src/epics` module definition (which amounts to basically being a
shared mutable global value). Following the established pattern for
the logger and store, you now have a factory method
`configureEpics()` which ensures a new instance every time. Now every
`Root` object gets its own epics that can live and die with it. This
is particularly important for testing, where there might be thousands
of roots.

Next, a function `addEpic(key, epic)` is added to the context, which will
conditionally install an epic matching `key` if it has not been
installed already. This follows the pattern of `addReducer`.