Use <div> instead of <span> when dynamically generating module routes and containers
The `getModuleRoutes` method appears to generate a routing entry for all loaded stripes modules.  It then maps those routes to a container node that sort of functions as a namespace for the different UI modules installed in the Folio instance.
We were using spans here, but it's probably better to use divs to avoid nesting block-level elements inside inline elements.  This isn't blocking anything that we know of, but we stumbled across it while doing some layout work and figured it was worthy of a quick patch.