Fix promise catch order
When you catch a promise, it does not stop the propagation of a
chain. This allows you to handle errors before continuing the chain.

So when the discovery promises were failing and getting caught, the
following `.then` chains were still called, except with the argument
returned from the catch statement. Since the catch statements do not
return anything, the resulting `response` is `undefined`.

This simply moves those catch statements to be at the end of their
respective chains so when a failure occurs it does not continue. If
the promise is rejected before a preceding `then` block, that block
will be skipped and the promise will immediately call the first
`catch` statement it sees.