Label a tenant's disabled modules on `/about`. Fixes STCOR-69
In order to label a tenant's disabled modules, we need to be able to
query Okapi from within stripes-core, but `stripes.connect()`, which is
actually the result of calling `connectFor('@folio/some-module')` for
each dynamically-loaded module, wasn't available within stripes-core.

The approach taken here was to hard-code a call to
`connectFor('@folio/core')` and stash it in the context. The benefit of
using context is that we stash it in one place and every component that
wants it just specifies a static `contextTypes = { coreConnect:
PropTypes.func,}` element. The downside is that connecting a component
in stripes-core is now slightly different than connecting a component in
an app even though the end result is the same.

RootWithIntl had to be refactored into a stateful component because
of how `getChildContext` is implemented; `context` is not avaiable in
child components if the context is defined on a stateless component.