Label a tenant's disabled modules on `/about`. Fixes STCOR-69
In order to label a tenant's disabled modules, we need to be able to
query Okapi from within stripes-core, but `stripes.connect()`, which is
actually the result of calling `connectFor('@folio/some-module')` for
each dynamically-loaded module, wasn't available within stripes-core.

The approach taken here was to hard-code a call to
`connectFor('@folio/core')` on `RootWithIntl` and to replace the dummy
version of `stripes.connect` with the result of that call, as is done in
`moduleRoutes.js::getModuleRoutes()` for dynamically loaded modules.
This is in contrast to the approach taken in PR #81 where `connect` was
placed in the context rather than directly on the cloned stripes object.