Add switchMap operator
When modules such as `eholdings` are loaded into a folio application
the epic they recieve comes from stripes-core. So if you need access
to a method/rxjs operator off of `action$` such as
`action$.ofType().switchMap()` in that module it needs to be
in stripes-core so when stripes-core combines its epics
the operator will be present.