DRY out notification text parsing. Fixes STCOR-104
Use a single constant for the regular expression responsible for finding
UUIDs in the text, and import a function to handle the mapping from a
note's link's domain attribute to a URL so that it can easily be
replaced with something that reads from an application registry once we
finally write an application registry.

You may note that there are many UUID regular expression packages out
there in NPM land. All of them return an actual RegEx object, but in
this case we need the string that we can construct our own RegEx object
around.