<Login> now accepts "authError" (string) property
This replaces the old "authFail" boolean. In the past, `<Login>` would
generate its own human-readable error-message when there was an error,
but this made it incapable of distinguishing between different kinds
of errors -- for example, a back-end internal error was reported the
same as a username/password mismatch: see STCOR-119.

Now it is the responsibility of the caller to generate and pass in an
informative message.