When server is absent, say so instead of displaying nothing.
This turned out to be much more than the trivial fix I'd imagined,
since Stripes was making no distinction between "I don't have a
response to my first call _yet_" and "My first call failed". I had to
add a new bit of state, `serverDown`, to the Redux store – which, as
always, involved touching a whole bunch of files: `okapiActions.js` to
define the `setServerDown` function that makes an action,
`okapiReducer.js` to recognise that action and set the bit of state,
`loginServices.js` to invoke that action when the initial call fails,
and finally `Root.js` to emit the error message when this has
failed. Oh, and the `CHANGELOG.md` of course.

Fixes STCOR-164.