refactoring to appease eslint
I'm torn. ESLint bans "continue" statements because they make code
more difficult to read "when used incorrectly," which may be true, but
not every use is "incorrect" and not every use is unclear. The change
here was from throwing an exception to break control flow to using a
continue statement to break control flow; if the former was acceptable
there is no reason the latter shouldn't be.

That said, I generally favor "if (test) action; else warning;"
over "if (! test) warning; else action" and the lint complaint resulted
in this structure. So, here we are, doing the right thing for the wrong
reasons.