STCOR-196: Drop translation strings that are no longer in use.
Translations keys that are no longer in use should be deleted so that
the translator don't waste their time.

In addition translation keys that lack context are difficult to translate.

These 5 keys are not used in any https://github.com/folio-org
repository, I provide the search link for each of them:

* button.close https://github.com/search?utf8=%E2%9C%93&q=org%3Afolio-org+%22button.close%22&type=Code
* button.saveChanges https://github.com/search?utf8=%E2%9C%93&q=org%3Afolio-org+%22button.saveChanges%22&type=Code
* button.discardChanges https://github.com/search?q=org%3Afolio-org+%22button.discardChanges%22&type=Code
* button.keepEditing https://github.com/search?utf8=%E2%9C%93&q=org%3Afolio-org+%22button.keepEditing%22&type=Code
* label.unsavedChanges https://github.com/search?utf8=%E2%9C%93&q=org%3Afolio-org+%22label.unsavedChanges%22&type=Code