keep stripes object up to date after login
The work done in support of STCOR-273 refactored how the `stripes`
object available in `render()` was initialized. By moving the
initialization and cloning up to the constructor, it wasn't updated when
user permissions, among other attributes, were added to the object.
Previously, because the object was cloned in `render()`, these
attributes were always available, but relocating the clone operation to
the constructor means these updates never trickled down.

This problem was captured in STRIPES-578 and patched, but there were
lingering issues as noted in STCOR-283. So _this_ refactor restores the
clone operation to `render()`, ensuring that all and any important
attributes, not just those called out in `componentDidUpdate()`, trickle
down to the `stripes` object used there.

It works by moving the `stripes.connect()` operation for
`<CreateResetPassword>` out of `RootWithIntl` and into a wrapper object
which receives the `stripes` object configured here. That allows
`<CreateResetPassword>` to get connected without restructuring
`<RootWithIntl>`.

Refs STCOR-273, STRIPES-578, STCOR-283