load region-specific translations if available
react-intl expects locales to look like `pt-BR`, i.e. a locale, a hyphen,
and a region code. lokalise.com provides translation files with names
like `pt_BR.json`, i.e. a locale, an underscore, and a region code. In
other words, the two systems we use for managing translations DON'T KNOW
HOW TO TALK TO EACH OTHER. For reals.

Regardless, the old code threw away region codes altogether, so it
didn't really matter that they couldn't talk because they weren't even
trying. Now, we translate the dash to an underscore in order to look up
the locale + region and use that if available, and fall back to the
locale-only, which is what we used to do, otherwise.

Fixes STCOR-261