provide default translations from en.json
Include `en.json` as the base for all translations so that keys missing
from a locale file but present in the `en.json` file will show the value
from the `en.json` file rather than the key. This will allow developers
to add pairs to the `en.json` file and have the values immediately show
up in the UI regardless of locale.

And this time, do it without merging all translations into those English
translations, thus causing the values from the last locale in the
iteration to overwrite the values from locales. Ahem. #559. * cough *

Fixes STCOR-310