Allow modules to be of more than one type. (STCOR-148) (#684)
Allow modules to be ingested by stripes-core as more than one type -- the type property of the stripes section is now replaced by actsAs which can be a module type string or an array of them. Also removes redundant copies of the icons, Okapi interfaces, and permissions in the modules.<type> structure exposed by the ModulesContext as they are kept elsewhere.