UIIN-678 receive react via peerDependencies
We should only provide react in one place (stripes) and every other
library and app built with stripes should recieve it as a peer. This
reduces the chance of different modules using different minor version of
react with a `~` dependency and pulling multiple versions of it into the
bundle. Multiple versions of React === Bad Things.

Refs UIIN-678