Allow top-left "FOLIO" caption to be configured
Fixes STCOR-385.

If the `stripes.congig.js` file's `config` section contains
`platformName` or `platformDescription`, then the corresponding values
will be used for the name and description of the top-level caption.

If these values are not provided, the code defaults to the old
behaviour of using a hardwired "FOLIO" caption and "FOLIO Platform"
description (i.e. this change is backwards-compatible).