STCOM-581, STCOR-390 react-router updates
Update react-router and react-router-dom to v5 and move them from
dependencies to peerDependencies in order to guarantee that all modules
in a platform are on the same version. (STCOM-581)

Updating from v4 to v5 causes warnings in a few components that ask for
`router` in the context but find it to be undefined. In fact, in every
instance where `router` is explicitly pulled from the context, we don't
actually use it. I think it is likely just left over from previous
incarnations where we called `...router.history.push` and it was never
removed when that code was refactored away.

Refs STCOM-581, STCOR-390