STCOR-393 release a later "latest" stripes-core
Jenkins wasn't fooled when we simply added an additional tag without any
other changes; it wouldn't run the pipeline on the tag to completion. So
now we'll make an _actual_, though technically ineffectual, change to
the project to hopefully trick Jenkins _and_ yarn into seeing this as an
update worthy of the `latest` `dist-tag`. (Ineffectual because changing
from `^4.0.0` to `^4.3.1` when `v4.3.1` has been out for a year won't
change the package that should be installed.)

Refs STCOR-393