Add explicit dependency on handlebars
I previously assumed this was not necessary, as the dependency on the
WebPack loaded `handlebars-loader` would transitively pull in
`handlebars` itself.

It turns out that the loader has the templating engine only as a
peer-dependency, not a full dependency, which means that building
stripes-core in the absence of a module that has its own full
dependency on `handlebars` will fail due to the unfulfilled
peer-dependency.

This will fix a persistent and mysterious build problem first
encountered in ReShare, but which is sure to crop up again in other
contexts if not addressed.