Cherry pick the FakeXMLHttpRequest fix from #761 (#763)
Cherry pick the `FakeXMLHttpRequest` fix from #761 onto the `b3.10` branch.

This ugly PR patches `FakeXMLHttpRequest` based on the discoveries from @VictorSoroka1 in 

https://github.com/pretenderjs/FakeXMLHttpRequest/issues/54

which were introduced by:

https://github.com/pretenderjs/FakeXMLHttpRequest/pull/53/files

The changes introduced here together with the PR from @VictorSoroka1  in:

https://github.com/folio-org/stripes-smart-components/pull/677

should hopefully resolve the test issues. 

This is a temporary fix until we can get the Mirage dependency updated.