STCOR-415 replace deprecated webapp-webpack-plugin (#798)
We replaced `favicons-webpack-plugin` with `webapp-webpack-plugin`
because the former had a memory leak, but the latter is now deprecated
and has been folded back into the former.

This requires that we update to a compatible version of 
`html-webpack-plugin`, i.e. `^4.0.0-beta` as well. On the one hand I don't
like the `-beta`, on the other, it's been in beta since August, 2018. That
feels more like a permanent-beta of the "Gmail beta" sense and I don't 
mind that at all. 

Refs STCOR-415, STCOR-232, STCOR-296

Co-authored-by: Matt Jones <mattjones@ebsco.com>