show app icons in settings navigation (#868)
Nobody asked for this but I think its helpful. Say, _totally
hypothetically_, you are browsing folio and the locale has been set to
an unfamiliar language and you would like to change it back. Wouldn't it
be convenient if you could see the app icons so you don't have to mouse
over every single option and look at the URL in order to find
`settings/tenant-settings`?

At least, that was the promise. It turns out the settings-only apps
universally don't provide app-icons, so they all show the same empty
grey box for an icon. **sigh** I still think this is pretty though.

Co-authored-by: Rasmus Wølk <rasmuswoelk@gmail.com>
Co-authored-by: Rasmus Wølk <640976+rasmuswoelk@users.noreply.github.com>