STCOM-452 actually validate the stored token
Previous implementations tested the validity of a stored token by using
to conduct a dummy fetch. This was problematic for two reasons:

* It was an extra request and we just threw away the response as long as
it was a 200.
* Even if the token was valid, the request would still fail if the user
retrieved from the token didn't have permission on that endpoint.

Now, we parse the token and validate its shape, which is what this
function was supposed to be doing all along.

Refs STCOR-452