STCOR-452 validate token without permissions
The `validateUser()` function actually does three things:

* check if the given token is syntactically valid
* check whether Okapi is up
* check if the token can be used to make a request

It does this in one fell swoop by making a request, but in order for
this to function for _all_ users the endpoint it hits must be one that
does not require any permissions. `/users` is not such an endpoint, but
`/bl-users/_self` is.

Replaces #923, where I attempted to resolve the permissions problem by
checking for token validity but, ahem, also ignoring the other two tasks
this function accomplishes.

Refs STCOM-452