STCOR-496 conduct discovery after login
Backport changes from STCOR-454 (PR #893, #901) to `v5` in order to
allow operators in a multi-tenant environment to update their okapi to
v4. This means conducting module-discovery _after_ login and
authenticated as the current tenant instead of `supertenant`.

In short, this change allows `@folio/stripes` `v4`
(`@folio/stripes-core` `v5`) to be compatible with `okapi` `v4`;
previously, it was compatible only with `v3`.

Refs STCOR-496, STCOR-454