STCOR-504 do not access empty object
folio-testing shows an empty screen with the following in the console:
```
Uncaught TypeError: Cannot read property 'isLocalLogin' of undefined
    at ProfileDropdown.value (bundle.310dcdca34d1f4011c62.js:483)
    at bundle.310dcdca34d1f4011c62.js:483
    at Array.map (<anonymous>)
    ...
```
I can't explain why `module.getModule()` is being called in a way that
returns empty, which results in failing to be able to access
`isLocalLogin`, but that's what's happening.

Truly, I'm stumped about the root cause here, and this is clearly a
bandaid solution, but at least it restores access, hopefully, to
folio-testing.

Refs STCOR-504