STCOR-455 publish ERROR events from RouteErrorBoundary
Publish `ERROR` events from `<RouteErrorBoundary>` and provide a
callback to the `<ErrorBoundary>` component that allows its
`componentDidCatch` method to provide error details to the modules that
have registered to handle these `ERROR` events. (Modules of type
`handler` are basically subscribers to events published by stripes-core.)

While existing event-handlers return components to interrupt the render
process, error logging is a bit different: we _don't_ want to interrupt
that process; we just want to log the error. Hence the renaming of
methods such as `getHandlerComponents` to `getEventHandlers` since the
handlers may be components, or they may simply be functions.

Refs STCOR-455, STCOR-493, STCOM-753