STCOR-514 refactor to avoid context-related error
This maybe broke when we upgraded `react-redux` in #985, though
possibly it was broken earlier since the correct context for statically
configuring context is `contextType` and here we had `contextTypes`, but
it's been like that for years, so who knows.

In any case, the reported bug was unreproduceable, but investigating did
lead me to this refactoring, so I'm not unhappy.

Refs STCOR-514