FOLIO-3253 tweak login to avoid users-bl
POC refactor of login to avoid `users-bl`, which has lots of
dependencies, in favor of `authn/login`, which has very few. This is not
intended to be pretty (obviously, the extra requests need cleaning up
and error handling) but just to demonstrate that it is possible to login
from the UI against a minimal set of back-end and front-end modules.

Refs FOLIO-3253