clean up, clean up, everybody clean up!
Special thanks to @MikeTaylor for pointing out that trading a clean
lint-run for a metric shittonne of dev-deps didn't really seem like a
great trade.

The initial commit added many dev-deps here that are, in fact, managed
wholly by `stripes-webpack` and therefore do not need to be included in
repositories that already rely on `stripes-webpack` (spoiler: all FOLIO
UI repositories already rely on `stripes-webpack`, which is a hard-dep
of `stripes-cli`).