STCOR-480 STCOR-525 restore okapi version retrieval
`/_/version` was available, and then it wasn't, and now it is. Now
shake it all about, do the hokey pokey and turn yourself around, if you
aren't already too turned around.

This started merely as an attempt to deal with a warning about a bad
prop-types setting (`PropTypes.okapi` is not a thing) but as I got in
there and started looking around to figure out the correct prop-type,
and then to add a comment referencing the ticket when the okapi-version
info was removed, I discovered it had been restored.

Sorted the stripes-shape prop-types for good measure.

Refs STCOR-480, STCOR-525, OKAPI-917.