STCOR-590 align prop-types for password reset errors (#1166)
There was general disagreement about the expected prop-types for errors
related to password reset. Generally, more things seem to expect an
array than an object, leading me to think the handling of the one-error
case, where the error is directly dispatched, was just a mistake. At
least, converting that one case resolved a litany of errors, and then
the only downstream change was fixing its message display, which, come
to think of it, was probably already broken and we just never noticed
and always accepted the default error message.

Refs STCOR-590