Update ky
Noticed we were several versions behind, was curious to see what changed
so I checked. Nothing that would break our use of it (though I did
search the folio org for beforeRetry as that change in .24 could
possibly... we don't seem to use it though).

In addition to some various improvements (especially error handling) and
tidying along the way, the bugfix in .27 to prevent a spurious parameter
from being added seems quite worthwhile indeed.