assert on dynamically applied style, not static style attributes
Previous test asserted on the rendered size of an image, e.g.
`size==14`, which is a static attribute of "small" images. That's not
super valuable; it's essentially asking "Does the CSS class contain the
value `14`?" when the question we really want to ask is, "Was the
correct CSS class applied?"