STCOR-620 separate route- and component-based tests (#1193)
Route-based tests that leverage BigTest to mount and component-based
tests that mount directly don't play nice together. The problem appears
to be with how the DOM node under which a component is mounted is
cleaned up, or not, and a fight for control of this node ensues. For
reasons unclear, even if both native code and BigTest are given the same
`id` value, they don't use the same element; instead they each create
their own element with the same `id`.

This PR handles this in two ways:

1. Leverage `@bigtest/react`'s `setupAppForTesting` function to handle
the mounting/unmounting of all components for all tests, including both
route-based tests that require mounting the entire application and
component-based tests that mount only a single component.

2. Run all the component-based tests first (these are present in 
`src/components/...`) followed by all the route-based tests 
(these are present in `test/bigtest/tests/...`). Hopefully, running 
strictly in series prevents them from stepping on each other's toes.

Refs STCOR-620