restore @bigtest/mirage
`@bigtest/mirage` was removed in folio-org/stripes-webpack/pull/61
because it is not invoked directly by `stripes-webpack`; rather it
should be a dev-dep only of the modules that need it for testing.
Probably it was moved from stripes-core to stripes-webpack incorrectly
when we split the two, and we missed it during earlier cleanup efforts,
e.g. #1146.