STCOR-621 point MainNav to documentation's root URL
Point to the documentation site's root URL, https://docs.folio.org/,
instead of https://docs.folio.org/docs/, which corresponds to the
Honeysuckle release. This will have the effect of keeping the destinaton
link "current" since the "Go to the docs" button at the root URL always
points to the most-recent release.

Refs STCOR-621