Mending Wall
Something there is that doesn't comment,
That writes the functions higgeldy piggeldy,
And nests the callbacks and the lookbacks;
And makes a rat's nest even a rat wouldn't call home.
The work of bug-patchers is another thing:
I have come after them and made repair
Where they have left not one brief note about their change,
But they would commit it PR-less straight to master
To please the yelping clients.